change data type

+++ db/migrate/20160224025318_add_columns_to_posts.rb
@@ -1,6 +1,6 @@
 class AddColumnsToPosts < ActiveRecord::Migration
   def change
-  	add_column :posts, :price, :double
+  	add_column :posts, :price, :float
   	add_column :posts, :category, :string
   end
 end
